refactor: extract ListOptions, move app session getter back to service
